+++ blockchain/src/main/java/com/ecommerce/blockchain/service/WalletService.java
@@ -1,9 +1,5 @@
 package com.ecommerce.blockchain.service;
 
-import com.ecommerce.blockchain.domain.user.User;
-import com.ecommerce.blockchain.domain.wallet.WalletRegistReq;
-import com.ecommerce.blockchain.domain.wallet.WalletResponseDto;
-
 import java.io.IOException;
 import java.util.concurrent.ExecutionException;
 
@@ -16,6 +12,6 @@ public interface WalletService {
 //    WalletResponseDto getWalletById(User user) throws IOException;
 
 //    WalletResponseDto chargeEther(String address) throws IOException, ExecutionException, InterruptedException;
-boolean chargeNEToken(int chargeNEToken, String address) throws IOException, ExecutionException, InterruptedException;
+boolean chargeNEToken(Long chargeNEToken, String address) throws IOException, ExecutionException, InterruptedException;
 
 }
